WIP checkpoint: consolidations complete, DuckDB query layer in progress
Consolidations (complete, not yet compiled): shared AI assistant core in
ai-config consumed by both MCP modules; guardian guard pairs merged into
parameterized LoggingOverheadGuard/LoggingBlockingGuard/ZGarbageCollectionGuard;
single-impl interfaces collapsed (RemoteServersManager, RemoteServerManager,
ProjectRecordingInitializer, CollapseFramesManager); generic DuckDB appender
null-helpers centralized in shared/persistence.

DuckDB query-layer rework still mid-flight in this snapshot - do not
review this commit in isolation; final verified commit follows.
