+++ jeffrey-microscope/core-microscope/src/main/java/cafe/jeffrey/microscope/core/recording/ProjectRecordingInitializer.java
@@ -18,31 +18,51 @@
 
 package cafe.jeffrey.microscope.core.recording;
 
+import cafe.jeffrey.shared.common.IDGenerator;
+import cafe.jeffrey.shared.common.filesystem.FileSystemUtils;
 import cafe.jeffrey.shared.common.model.ProjectInfo;
+import cafe.jeffrey.shared.common.model.Recording;
+import cafe.jeffrey.shared.common.model.RecordingFile;
 import cafe.jeffrey.shared.common.model.repository.RepositoryFile;
+import cafe.jeffrey.shared.common.model.repository.SupportedRecordingFile;
 import cafe.jeffrey.microscope.core.persistence.NewRecordingHolder;
+import cafe.jeffrey.provider.profile.api.RecordingInformationParser;
 import cafe.jeffrey.microscope.core.persistence.NewRecording;
+import cafe.jeffrey.provider.profile.api.RecordingInformation;
+import cafe.jeffrey.microscope.persistence.api.RecordingRepository;
 import cafe.jeffrey.storage.recording.api.ProjectRecordingStorage;
 
 import java.nio.file.Path;
+import java.time.Clock;
+import java.time.Instant;
 import java.util.List;
 import java.util.function.Function;
 
-public interface ProjectRecordingInitializer {
+public class ProjectRecordingInitializer {
 
     @FunctionalInterface
-    interface Factory extends Function<ProjectInfo, ProjectRecordingInitializer> {
+    public interface Factory extends Function<ProjectInfo, ProjectRecordingInitializer> {
     }
 
-    /**
-     * Initializes a new recording and provides a {@link NewRecordingHolder} that contains resources
-     * for managing the recording's lifecycle. It allows copying some additional files to the recording as well.
-     *
-     * @param recording       the details of the recording including filename, group ID, and input stream
-     * @param additionalFiles a list of additional files to be associated with the recording
-     * @return a {@link NewRecordingHolder} that provides access to the output stream for uploading recording files.
-     */
-    NewRecordingHolder newRecording(NewRecording recording, List<RepositoryFile> additionalFiles);
+    private final RecordingRepository recordingRepository;
+    private final ProjectRecordingStorage recordingStorage;
+    private final RecordingInformationParser recordingInformationParser;
+    private final Clock clock;
+    private final ProjectInfo projectInfo;
+
+    public ProjectRecordingInitializer(
+            Clock clock,
+            ProjectInfo projectInfo,
+            ProjectRecordingStorage recordingStorage,
+            RecordingRepository recordingRepository,
+            RecordingInformationParser recordingInformationParser) {
+
+        this.clock = clock;
+        this.projectInfo = projectInfo;
+        this.recordingStorage = recordingStorage;
+        this.recordingInformationParser = recordingInformationParser;
+        this.recordingRepository = recordingRepository;
+    }
 
     /**
      * Creates a new recording and provides a {@link NewRecordingHolder} that contains resources
@@ -52,24 +72,178 @@ public interface ProjectRecordingInitializer {
      * @param recording the details of the recording including filename, group ID, and input stream
      * @return a {@link NewRecordingHolder} that provides access to the output stream for uploading recording files.
      */
-    default NewRecordingHolder newRecording(NewRecording recording) {
+    public NewRecordingHolder newRecording(NewRecording recording) {
         return newRecording(recording, List.of());
+    }
+
+    /**
+     * Initializes a new recording and provides a {@link NewRecordingHolder} that contains resources
+     * for managing the recording's lifecycle. It allows copying some additional files to the recording as well.
+     *
+     * @param newRecording    the details of the recording including filename, group ID, and input stream
+     * @param additionalFiles a list of additional files to be associated with the recording
+     * @return a {@link NewRecordingHolder} that provides access to the output stream for uploading recording files.
+     */
+    public NewRecordingHolder newRecording(NewRecording newRecording, List<RepositoryFile> additionalFiles) {
+        String recordingId = IDGenerator.generate();
+        Path targetPath = recordingStorage.uploadTarget(recordingId, newRecording.filename());
+
+        Runnable uploadCompleteCallback = () -> {
+            if (newRecording.groupId() != null) {
+                boolean groupExists = recordingRepository.groupExists(newRecording.groupId());
+                if (!groupExists) {
+                    throw new RuntimeException("Group does not exist: " + newRecording.groupId());
+                }
+            }
+
+            try {
+                // Provide information about the Recording file
+                RecordingInformation information = recordingInformationParser.provide(targetPath);
+                validateRecordingInformation(information, newRecording.filename());
+
+                Instant createdAt = clock.instant();
+                Recording recording = new Recording(
+                        recordingId,
+                        newRecording.recordingName(),
+                        projectInfo.id(),
+                        newRecording.groupId(),
+                        information.eventSource(),
+                        createdAt,
+                        information.recordingStartedAt(),
+                        information.recordingFinishedAt(),
+                        false, null, null,
+                        List.of());
+
+                RecordingFile recordingFile = new RecordingFile(
+                        IDGenerator.generate(),
+                        recordingId,
+                        newRecording.filename(),
+                        SupportedRecordingFile.of(newRecording.filename()),
+                        createdAt,
+                        FileSystemUtils.size(targetPath));
+
+                recordingRepository.insertRecording(recording, recordingFile);
+
+                // ------------------------------------------------------
+                // Upload Additional files to the newly created recording
+                // ------------------------------------------------------
+
+                List<Path> additionalFilePaths = additionalFiles.stream()
+                        .map(RepositoryFile::filePath)
+                        .toList();
+
+                recordingStorage.addArtifacts(recordingId, additionalFilePaths);
+
+                for (RepositoryFile additionalFile : additionalFiles) {
+                    RecordingFile additionalRecordingFile = new RecordingFile(
+                            IDGenerator.generate(),
+                            recordingId,
+                            additionalFile.name(),
+                            additionalFile.fileType(),
+                            createdAt,
+                            additionalFile.size());
+
+                    recordingRepository.insertRecordingFile(additionalRecordingFile);
+                }
+
+            } catch (Exception e) {
+                FileSystemUtils.removeFile(targetPath);
+                throw new RuntimeException("Failed to upload recording: " + newRecording.filename(), e);
+            }
+        };
+
+        return new NewRecordingHolder(recordingId, targetPath, uploadCompleteCallback);
     }
 
     /**
      * Initializes a new recording and provides a {@link NewRecordingHolder} that contains resources
      * for managing the recording's lifecycle. It allows copying additional files (by path) to the recording.
      *
-     * @param recording           the details of the recording including filename, group ID, and input stream
+     * @param newRecording        the details of the recording including filename, group ID, and input stream
      * @param additionalFilePaths paths to additional files to be associated with the recording
      * @return a {@link NewRecordingHolder} that provides access to the output stream for uploading recording files.
      */
-    NewRecordingHolder newRecordingWithPaths(NewRecording recording, List<Path> additionalFilePaths);
+    public NewRecordingHolder newRecordingWithPaths(NewRecording newRecording, List<Path> additionalFilePaths) {
+        String recordingId = IDGenerator.generate();
+        Path targetPath = recordingStorage.uploadTarget(recordingId, newRecording.filename());
+
+        Runnable uploadCompleteCallback = () -> {
+            if (newRecording.groupId() != null) {
+                boolean groupExists = recordingRepository.groupExists(newRecording.groupId());
+                if (!groupExists) {
+                    throw new RuntimeException("Group does not exist: " + newRecording.groupId());
+                }
+            }
+
+            try {
+                // Provide information about the Recording file
+                RecordingInformation information = recordingInformationParser.provide(targetPath);
+                validateRecordingInformation(information, newRecording.filename());
+
+                Instant createdAt = clock.instant();
+                Recording recording = new Recording(
+                        recordingId,
+                        newRecording.recordingName(),
+                        projectInfo.id(),
+                        newRecording.groupId(),
+                        information.eventSource(),
+                        createdAt,
+                        information.recordingStartedAt(),
+                        information.recordingFinishedAt(),
+                        false, null, null,
+                        List.of());
+
+                RecordingFile recordingFile = new RecordingFile(
+                        IDGenerator.generate(),
+                        recordingId,
+                        newRecording.filename(),
+                        SupportedRecordingFile.of(newRecording.filename()),
+                        createdAt,
+                        FileSystemUtils.size(targetPath));
+
+                recordingRepository.insertRecording(recording, recordingFile);
+
+                // ------------------------------------------------------
+                // Upload Additional files to the newly created recording
+                // ------------------------------------------------------
+
+                recordingStorage.addArtifacts(recordingId, additionalFilePaths);
+
+                for (Path additionalFilePath : additionalFilePaths) {
+                    String filename = additionalFilePath.getFileName().toString();
+                    RecordingFile additionalRecordingFile = new RecordingFile(
+                            IDGenerator.generate(),
+                            recordingId,
+                            filename,
+                            SupportedRecordingFile.of(filename),
+                            createdAt,
+                            FileSystemUtils.size(additionalFilePath));
+
+                    recordingRepository.insertRecordingFile(additionalRecordingFile);
+                }
+
+            } catch (Exception e) {
+                FileSystemUtils.removeFile(targetPath);
+                throw new RuntimeException("Failed to upload recording: " + newRecording.filename(), e);
+            }
+        };
+
+        return new NewRecordingHolder(recordingId, targetPath, uploadCompleteCallback);
+    }
 
     /**
      * Returns the recording storage for accessing recording files.
      *
      * @return the project recording storage
      */
-    ProjectRecordingStorage recordingStorage();
+    public ProjectRecordingStorage recordingStorage() {
+        return recordingStorage;
+    }
+
+    private static void validateRecordingInformation(RecordingInformation information, String filename) {
+        if (information.recordingStartedAt() == null || information.recordingFinishedAt() == null) {
+            throw new RuntimeException(
+                    "Recording file is empty or corrupt (contains no JFR chunks): " + filename);
+        }
+    }
 }
