+++ jeffrey-microscope/profiles/profile-sql-persistence/src/main/java/cafe/jeffrey/provider/profile/jdbc/JdbcProfileEventTypeRepository.java
@@ -95,6 +95,9 @@ public class JdbcProfileEventTypeRepository implements ProfileEventTypeRepositor
     private static final String COLUMNS_BY_SINGLE_EVENT =
             "SELECT columns FROM event_types WHERE name = (:code) LIMIT 1";
 
+    private static final String PLACEHOLDER_SPAN_FILTER = "<<span_filter>>";
+    private static final String EVENTS_TABLE = "events";
+
     //language=SQL
     private static final String EVENT_SUMMARIES_BY_CODES = """
             WITH aggregated_events AS (
@@ -104,15 +107,7 @@ public class JdbcProfileEventTypeRepository implements ProfileEventTypeRepositor
                     SUM(weight) as weight
                 FROM events
                 WHERE event_type IN (:codes)
-                    AND (:span_filter_enabled = FALSE OR EXISTS (
-                        SELECT 1 FROM (
-                            SELECT UNNEST([:span_thread_hashes]) AS th,
-                                   UNNEST([:span_from_ms]) AS f,
-                                   UNNEST([:span_to_ms]) AS t
-                        ) iv
-                        WHERE events.thread_hash = iv.th
-                          AND EPOCH_MS(events.start_timestamp) BETWEEN iv.f AND iv.t
-                    ))
+                    <<span_filter>>
                 GROUP BY event_type
             )
             SELECT
@@ -246,7 +241,7 @@ public class JdbcProfileEventTypeRepository implements ProfileEventTypeRepositor
         SpanIntervalParams.apply(paramSource, null);
 
         List<EventSummary> eventSummaries = databaseClient.query(
-                StatementLabel.EVENT_SUMMARIES, EVENT_SUMMARIES_BY_CODES, paramSource, EVENT_SUMMARY_MAPPER);
+                StatementLabel.EVENT_SUMMARIES, eventSummariesByCodesSql(null), paramSource, EVENT_SUMMARY_MAPPER);
 
         List<EventSummary> calculatedEventSummaries = commonCalculators.stream()
                 .filter(calculator -> types.contains(calculator.type()))
@@ -269,7 +264,18 @@ public class JdbcProfileEventTypeRepository implements ProfileEventTypeRepositor
 
         // Calculated (non-DB) summaries are profile-wide and cannot be span-scoped, so they are omitted here.
         return databaseClient.query(
-                StatementLabel.EVENT_SUMMARIES, EVENT_SUMMARIES_BY_CODES, paramSource, EVENT_SUMMARY_MAPPER);
+                StatementLabel.EVENT_SUMMARIES, eventSummariesByCodesSql(spanIntervals), paramSource, EVENT_SUMMARY_MAPPER);
+    }
+
+    /**
+     * Splices the span-scope semi-join into the summaries query only when span intervals are present;
+     * without span scoping, the clause is absent entirely.
+     */
+    private static String eventSummariesByCodesSql(List<SpanInterval> spanIntervals) {
+        String spanFilter = SpanIntervalParams.enabled(spanIntervals)
+                ? SpanIntervalParams.semiJoinFragment(EVENTS_TABLE)
+                : "";
+        return EVENT_SUMMARIES_BY_CODES.replace(PLACEHOLDER_SPAN_FILTER, spanFilter);
     }
 
     @Override
