+++ jeffrey-microscope/profiles/profile-sql-persistence/src/test/resources/sql/events/insert-spans.sql
@@ -11,11 +11,12 @@ VALUES
     (2001, 'http-nio-exec-3', 41, 12, false),
     (2002, 'hprof-indexer', 71, 33, false);
 
--- Durations in nanoseconds. First event at 10:00:00.000 establishes the relative-time origin.
-INSERT INTO events (event_type, start_timestamp, duration, samples, weight, weight_entity, stacktrace_hash, thread_hash, fields)
+-- Durations in nanoseconds. The profiling start (10:00:00.000) is the relative-time origin
+-- persisted in start_timestamp_from_beginning at ingest time.
+INSERT INTO events (event_type, start_timestamp, start_timestamp_from_beginning, duration, samples, weight, weight_entity, stacktrace_hash, thread_hash, fields)
 VALUES
-    ('profiler.Span',       '2025-01-15T10:00:00.000Z', 300000000, 1, NULL, NULL, NULL, 2001, '{"tag":"profile.initialize"}'),
-    ('profiler.Span',       '2025-01-15T10:00:00.100Z', 120000000, 1, NULL, NULL, NULL, 2001, '{"tag":"jfr.parse_and_ingest"}'),
-    ('profiler.Span',       '2025-01-15T10:00:02.000Z', 200000000, 1, NULL, NULL, NULL, 2002, '{"tag":"hprof.index.build"}'),
+    ('profiler.Span',       '2025-01-15T10:00:00.000Z',    0, 300000000, 1, NULL, NULL, NULL, 2001, '{"tag":"profile.initialize"}'),
+    ('profiler.Span',       '2025-01-15T10:00:00.100Z',  100, 120000000, 1, NULL, NULL, NULL, 2001, '{"tag":"jfr.parse_and_ingest"}'),
+    ('profiler.Span',       '2025-01-15T10:00:02.000Z', 2000, 200000000, 1, NULL, NULL, NULL, 2002, '{"tag":"hprof.index.build"}'),
     -- Non-span event: must NOT be returned by listSpans().
-    ('jdk.ExecutionSample', '2025-01-15T10:00:00.050Z',      NULL, 1, NULL, NULL, NULL, 2001, '{"state":"RUNNABLE"}');
+    ('jdk.ExecutionSample', '2025-01-15T10:00:00.050Z',   50,      NULL, 1, NULL, NULL, NULL, 2001, '{"state":"RUNNABLE"}');
